+++ bamazonCustomer.js
@@ -1,5 +1,6 @@
 var inquirer = require("inquirer");
 var mysql = require("mysql");
+const cTable = require('console.table');
 
 
 var connection = mysql.createConnection({
@@ -10,33 +11,60 @@ var connection = mysql.createConnection({
     database: "bamazon"
 });
 
-connection.connect(function(err, res){
-    if(err){
+connection.connect(function (err, res) {
+    if (err) {
         throw err;
     }
-    else{
-        //console.log("Database connection made!")
+    else {
     }
 })
 
+connection.query("SELECT * FROM product", function (err, result) {
+    if (err) {
+        throw err;
+    }
+    //TODO Make pretty
+    //console.log(result);
+    console.table(result);
 
-
-inquirer
-    .prompt([
+    var choices = [];
+    for(var i = 0; i < result.length; i++){
+        choices.push(result[i].item_id + "");
+    }
+    //console.log(choices);
+    inquirer.prompt([
         {
-            type: "input",
-            message: "This is the first question",
-            name: "first"
+            type: "list",
+            message: "What Item would you like to buy?",
+            name: "item_id",
+            choices: choices
         },
         {
             type: "input",
-            message: "This is the second question",
-            name: "second"
+            message: "How many would you like to buy?",
+            name: "quantity"
         }
-        /* Pass your questions in here */
-    ])
-    .then(function (answers) {
-        console.log("Bamazon Customer Loaded")
-        console.log(answers);
-    });
+    ]).then(function (answers) {
+        connection.query("SELECT * FROM product WHERE item_id = ?", [answers.item_id], function (err, result) {
+            if (err) {
+                throw err;
+            }
+            var quantity = result[0].stock_quantity;
+            var price = result[0].price;
+            var product_sales = result[0].product_sales;
+            if (answers.quantity > quantity) {
+                console.log("Insufficient quantity!")
+            }
 
+            else {
+                connection.query("UPDATE product SET ? WHERE item_id = ?", [{ stock_quantity: quantity - answers.quantity, product_sales: product_sales + (price * answers.quantity) }, answers.item_id], function (err, result) {
+                    if (err) {
+                        throw err;
+                    }
+                    console.log("Your purchase cost you: $" + (price * answers.quantity));
+
+                })
+            }
+        })
+    });
+});